IVEC/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.5
django-filter==23.2.0
pytest==7.2.2
pytest-django==4.5.2